backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
gunicorn>=23.0,<24.0
whitenoise>=6.7,<6.8

# NLP / ML
scikit-learn>=1.5,<1.6
sentence-transformers>=3.1,<3.2
pandas>=2.2,<2.3
numpy>=2.1,<2.2

# Background & cache (later)
celery>=5.4,<5.5
redis>=5.0,<6.0

# DB
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0

# API docs
drf-spectacular>=0.27,<0.28
